+++ example/clj/todopipeline/pipeline.clj
@@ -81,7 +81,9 @@
         ;; build-steps. :home-dir is the directory where LambdaCD will store all it's internal data that should be persisted
         ;; over time, such as the last seen revisions of various git-repositories, the build history and so on.
         config {:home-dir home-dir
-                :name "some pipeline"}
+                :name "some pipeline"
+                :ui-config {:expand-active-default true
+                            :expand-failures-default true}}
         ;; wiring everything together everything that's necessary for lambdacd to run
         pipeline (core/assemble-pipeline pipeline-def config)
         ;; the ring handler